requirements.txt
aiohttp==3.9.3
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.1
amqp==5.3.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.11.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.1
cachetools==5.3.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.3
dataclasses-json==0.6.4
distro==1.9.0
dnspython==2.6.1
email-validator==2.1.0.post1
exceptiongroup==1.2.2
fastapi==0.109.2
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
frozenlist==1.4.1
google-api-core==2.16.2
google-api-python-client==2.48.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
httpx-ws==0.5.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
kombu==5.4.2
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langsmith==0.0.86
loguru==0.7.2
lxml==5.1.0
makefun==1.15.2
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.11.1
packaging==23.2
passlib==1.7.4
password-strength==0.0.3.post2
pika==1.3.2
pillow==10.2.0
pluggy==1.4.0
prompt_toolkit==3.0.50
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pyfiglet==0.8.post1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.7
python-pptx==0.6.23
python-resize-image==1.1.20
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.20.1
SQLAlchemy==2.0.25
starlette==0.36.3
tenacity==8.2.3
termcolor==1.1.0
tomli==2.2.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.0.post1
uvloop==0.19.0
vine==5.1.0
watchdog==6.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
WTForms==3.1.2
XlsxWriter==3.1.9
yarl==1.9.4
